+++ api_gateway/src/main/kotlin/org/thechance/api_gateway/endpoints/CartRoute.kt
@@ -7,8 +7,12 @@ import io.ktor.server.auth.jwt.*
 import io.ktor.server.request.*
 import io.ktor.server.routing.*
 import org.koin.ktor.ext.inject
+import org.thechance.api_gateway.data.localizedMessages.LocalizedMessagesFactory
 import org.thechance.api_gateway.data.model.MealRequestDto
+import org.thechance.api_gateway.data.model.notification.NotificationDto
+import org.thechance.api_gateway.data.model.notification.NotificationSender
 import org.thechance.api_gateway.data.model.restaurant.CartDto
+import org.thechance.api_gateway.data.service.NotificationService
 import org.thechance.api_gateway.data.service.RestaurantService
 import org.thechance.api_gateway.endpoints.utils.authenticateWithRole
 import org.thechance.api_gateway.endpoints.utils.extractLocalizationHeader
@@ -19,6 +23,8 @@ import org.thechance.api_gateway.util.Role
 
 fun Route.cartRoutes() {
     val restaurantService: RestaurantService by inject()
+    val localizedMessagesFactory by inject<LocalizedMessagesFactory>()
+    val notificationService: NotificationService by inject()
 
     authenticateWithRole(Role.END_USER) {
         route("/cart") {
@@ -40,7 +46,7 @@ fun Route.cartRoutes() {
                 respondWithResult(HttpStatusCode.OK, result)
             }
 
-            put("/replace"){
+            put("/replace") {
                 val language = extractLocalizationHeader()
                 val tokenClaim = call.principal<JWTPrincipal>()
                 val userId = tokenClaim?.get(Claim.USER_ID).toString()
@@ -55,6 +61,19 @@ fun Route.cartRoutes() {
                 val userId = tokenClaim?.get(Claim.USER_ID).toString()
                 val result = restaurantService.orderCart(userId, language)
                 respondWithResult(HttpStatusCode.OK, result)
+
+                val notificationTitle = localizedMessagesFactory.createLocalizedMessages(language).newOrderTitle
+                val notificationBody = localizedMessagesFactory.createLocalizedMessages(language).newOrderBody
+                val restaurant = restaurantService.getRestaurantInfo(
+                    languageCode = language, restaurantId = result.restaurantId!!
+                )
+                val orderNotification = NotificationDto(
+                    userId = restaurant.ownerId,
+                    title = notificationTitle,
+                    body = notificationBody,
+                    sender = NotificationSender.UNDEFINED.code
+                )
+                notificationService.sendNotificationToUser(notificationDto = orderNotification, language)
             }
         }
     }
